requirements.txt
streamlit
openai
pandas
openpyxl
Pillow
python-docx
